+++ docs-demo/basic/row-cell-mouse-event/RowCellHoverSelect.vue
@@ -1,8 +1,9 @@
 <script lang="ts" setup>
-import { ref } from 'vue';
+import { ref, useTemplateRef } from 'vue';
 import StkTable from '../../StkTable.vue';
 import { StkTableColumn } from '../../../src/StkTable/index';
 import CheckItem from '../../components/CheckItem.vue';
+const stkTableRef = useTemplateRef('stkTableRef');
 
 const stripe = ref(true);
 const rowActive = ref(true);
@@ -25,26 +26,42 @@ const dataSource = ref([
     { name: `Lucy`, age: 22, address: `Guangzhou`, gender: 'female' },
     { name: `Lily`, age: 24, address: `Shenzhen`, gender: 'female' },
 ]);
+
+function setCurrentRow(rowKeyOrRow: string | undefined | any) {
+    stkTableRef.value?.setCurrentRow(rowKeyOrRow);
+}
+
+function setSelectedCell(row: any, col: StkTableColumn<any>) {
+    stkTableRef.value?.setSelectedCell(row, col);
+}
 </script>
 <template>
-    <CheckItem v-model="stripe" text="斑马纹(stripe)"></CheckItem>
+    <CheckItem v-model="stripe" text="stripe(斑马纹)"></CheckItem>
     <br />
-    <CheckItem v-model="rowActive" text="行选中状态(rowActive)"></CheckItem>
-    <CheckItem v-model="cellActive" text="单元格选中状态(cellActive)"></CheckItem>
+    <CheckItem v-model="rowActive" text="rowActive(行选中状态)"></CheckItem>
+    <CheckItem v-model="cellActive" text="cellActive(单元格选中状态)"></CheckItem>
     <br />
-    <CheckItem v-model="rowHover" text="行悬浮状态(rowHover)"></CheckItem>
-    <CheckItem v-model="cellHover" text="单元格悬浮状态(cellHover)"></CheckItem>
+    <CheckItem v-model="rowHover" text="rowHover(行悬浮状态)"></CheckItem>
+    <CheckItem v-model="cellHover" text="cellHover(单元格悬浮状态)"></CheckItem>
     <br />
     <CheckItem
         v-model="rowCurrentRevokable"
-        text="行选中状态是否可取消(rowCurrentRevokable)"
+        text="rowCurrentRevokable(行选中状态是否可取消)"
     ></CheckItem>
     <br />
     <CheckItem
         v-model="selectedCellRevokable"
-        text="单元格选中状态是否可取消(selectedCellRevokable)"
+        text="selectedCellRevokable(单元格选中状态是否可取消)"
     ></CheckItem>
+    <hr />
+    <button class="btn" @click="setCurrentRow('Jack')">setCurrentRow('Jack')</button>
+    <button class="btn" @click="setSelectedCell(dataSource[0], columns[1])">
+        setSelectedCell('Jack-age')
+    </button>
+
     <StkTable
+        ref="stkTableRef"
+        row-key="name"
         :stripe="stripe"
         :row-active="rowActive"
         :cell-active="cellActive"